Extract Device Flow hook into shared/auth
The web and mobile useDeviceFlow hooks had drifted into near-duplicates of
the same request/poll/cancel state machine, so a fix in one (e.g. the recent
cancellation guards) had to be ported by hand to the other.

Move the state machine into shared/auth/useDeviceFlow.ts and let each platform
wrapper inject the pieces that are actually platform-specific: the availability
resolver, the client ID, and a system-browser opener (window.open on web,
Linking.openURL on mobile). The shared module uses no DOM or React Native APIs,
per the shared/ boundary rule in CLAUDE.md.
